+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 # Changelog
 
 ## Unreleased
+- (feat) pages/governance 1.1.4 | Read blocked proposals from notion
 
 - (fix) dapp-58 | pages/dappstore | Update ecosystem section - show 8 dapps